Handle UsernameNotFoundException thrown from UaaUserDatabase
If a token is supplied that contains a GUID for a user that isn't
present in the database, a UsernameNotFoundException is thrown. The
exception isn't handled by the calling method and since it's runtime it
propogates, triggering a HttpClientErrorException in any Spring security
app using UAA as the OAuth2 provider.
Instead the user lookup is surrounded by a try catch which throws an
appropriate InvalidTokenException, if the user isn't found, which is
handled correctly.