Previously the tests could create a new emailChangeEmailService with the appropriate brand but this is now controlled by the message source.
Tests must ensure that the message source's brand is reset to pivotal
before each test since they can be executed out of order.
testBeginEmailChangeWithOssBrand switches to the oss brand but doesn't
switch it back which means tests assuming that the pivotal brand is
being used (due to the prior reset behaviour) may fail.